+++ modules/tinymce/src/themes/silver/main/ts/api/Options.ts
@@ -265,9 +265,8 @@ const register = (editor: Editor): void => {
 
   registerOption('resize', {
     processor: (value) => value === 'both' || Type.isBoolean(value),
-    // TODO: TINY-8288 - This should be set to `true` once the issue with the theme rendering too early is resolved
     // Editor resize doesn't work on touch devices at this stage
-    default: !Env.deviceType.isTouch() && !editor.hasPlugin('autoresize')
+    default: !Env.deviceType.isTouch()
   });
 };